requirements.txt
absl-py==2.3.1
addict==2.4.0
aiofiles==24.1.0
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
anyio==4.10.0
argcomplete==3.6.2
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
basicsr==1.4.2
bcrypt==4.3.0
beautifulsoup4==4.13.5
boto3==1.40.19
botocore==1.40.19
Brotli==1.1.0
cachetools==6.2.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
choreographer==1.0.10
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.0
cryptography==45.0.6
cycler==0.12.1
Cython==3.1.3
dacite==1.7.0
decorator==5.2.1
Deprecated==1.2.18
dill==0.4.0
dnspython==2.7.0
easydict==1.13
eval_type_backport==0.2.2
exceptiongroup==1.3.0
executing==2.2.1
facexlib==0.3.0
fiftyone==1.7.2
fiftyone-brain==0.21.2
fiftyone_db==1.3.0
filelock==3.13.1
filterpy==1.4.5
fire==0.7.1
flatbuffers==25.2.10
fonttools==4.59.2
fsspec==2024.6.1
ftfy==6.3.1
future==1.0.0
gfpgan==1.3.8
glob2==0.7
graphql-core==3.2.6
grpcio==1.74.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
humanize==4.13.0
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
inflate64==1.0.3
insightface==0.7.3
ipython==8.18.1
jedi==0.19.2
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.5.2
jsonlines==4.0.0
kaleido==1.0.0
kiwisolver==1.4.7
lazy_loader==0.4
lia-web==0.2.3
llvmlite==0.43.0
lmdb==1.7.3
logistro==1.1.0
Markdown==3.8.2
MarkupSafe==2.1.5
matplotlib==3.9.4
matplotlib-inline==0.1.7
ml_dtypes==0.5.3
mongoengine==0.29.1
motor==3.6.1
mpmath==1.3.0
multivolumefile==0.2.3
narwhals==2.2.0
networkx==3.2.1
numba==0.60.0
numpy==1.26.4
onnx==1.19.0
onnxruntime-gpu==1.16.2
opencv-python==4.11.0.86
opencv-python-headless==4.12.0.88
orjson==3.11.3
packaging==25.0
pandas==2.3.2
paramiko==3.5.1
parso==0.8.5
pillow==11.0.0
platformdirs==4.4.0
plotly==6.3.0
pprintpp==0.4.0
prettytable==3.16.0
priority==2.0.0
prompt_toolkit==3.0.52
protobuf==6.32.0
psutil==7.0.0
pure_eval==0.2.3
py7zr==1.0.0
pybboxes==0.1.6
pybcj==1.0.6
pycparser==2.22
pycryptodomex==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pydash==8.0.5
Pygments==2.19.2
pymongo==4.9.2
PyNaCl==1.5.0
pyparsing==3.2.3
pyppmd==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzstd==0.17.0
rarfile==4.2
realesrgan==0.3.0
regex==2025.7.34
requests==2.32.5
retrying==1.4.2
rtree==1.4.1
s3transfer==0.13.1
sahi==0.11.34
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
shapely==2.0.7
simplejson==3.20.1
simsimd==6.5.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
sse-starlette==0.10.3
sseclient-py==1.8.0
stack-data==0.6.3
starlette==0.47.3
strawberry-graphql==0.281.0
stringzilla==3.12.6
sympy==1.13.3
tabulate==0.9.0
taskgroup==0.2.2
tb-nightly==2.21.0a20250828
tensorboard-data-server==0.7.2
termcolor==3.1.0
terminaltables==3.1.10
texttable==1.7.0
threadpoolctl==3.6.0
tifffile==2024.8.30
tomli==2.2.1
torch==1.13.1
torchaudio==2.7.1+cu118
torchvision==0.14.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
universal-analytics-python3==1.1.1
urllib3==1.26.20
voxel51-eta==0.14.4
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.3
wsproto==1.2.0
xmltodict==0.14.2
yapf==0.43.0
zipp==3.23.0



scipy
